+++ src/app/(basic)/articles/_feature/zenn/fetchZennContents.ts
@@ -1,7 +1,7 @@
 import {
   GetZennContentsResult,
   ZENN_CONTENT_API_URL,
-} from '@/src/app/api/article/zenn/route'
+} from '@/src/app/api/article/route'
 
 export const fetchZennContents = async (): Promise<GetZennContentsResult> => {
   const response = await fetch(ZENN_CONTENT_API_URL)
